+++ admin/manage-food.php
@@ -109,7 +109,7 @@
                         <td><?php echo $active ?></td>
                         <td>
                             <a href="<?php echo SITEURL; ?>admin/update-food.php?id=<?php echo $id; ?>" class="btn btn-outline-success">Update</a>
-                            <a href="<?php echo SITEURL; ?>admin/delete-food.php?id=<?php echo $id; ?>&image_name=<?php echo $image_name ?>" class="btn btn-outline-danger">Delete</a>
+                            <a href="<?php echo SITEURL; ?>admin/delete-food.php?id=<?php echo $id; ?>&image_name=<?php echo $image_name ?>" onclick="confirmDeleteFood(event)" class="btn btn-outline-danger">Delete</a>
 
                         </td>
                     </tr>
@@ -132,6 +132,14 @@
     </div>
 </div>
 
+
+<script>
+        function confirmDeleteFood(event) {
+            if (!confirm("Are you sure you want to delete this food?")) {
+                event.preventDefault();
+            }
+        }
+    </script>
